xpath_content : add and use fetch_xpath_content()
The function fetch_xpath_content($link, $xpath) extract from $link the
content discribed by $xpath.

This function is then called to update $entry_content just before
insertion or update of a ttrss_entries.

This way, the content of the page refered by $entry_link is directly
available in ttrss, which is a turn around for some feeds which only
provide partial content. xpath is used to select a specific part of
entry_link.